requirements.txt
google-auth==1.6.3
requests==2.22.0
python-dateutil==2.8.0
pytz==2019.1
kickshaws @ git+https://github.com/wcmc-research-informatics/kickshaws/#egg=kickshaws
sqlsrvwrapper @ git+https://github.com/wcmc-research-informatics/sqlsrvwrapper/#egg=sqlsrvwrapper


